variadic_function-Fixed

+++ 0x10-variadic_functions/0-sum_them_all.c
@@ -10,7 +10,7 @@
 
 int sum_them_all(const unsigned int n, ...)
 {
-size_t sum, times;
+int sum, times;
 va_list nums;
 va_start(nums, n);
 if (n == 0)
